Remove first two lines from `itemdef.csv` files
These files contain definitions for what can be found in the respective
`data.csv` files. Their first two lines, however, do not conform to the
same number of columns or the same format as the rest of those files.

In the first line, we would find a title for the dataset in question. We
can also find that in the README for each dataset. The second line
contains the file that has the algorithm for each dataset. It's always
the same: `default.js`. Therefore, we think we can do without those two
lines.

This was done using the following sh-fu:
```sh
find . -name itemdef.csv | xargs sed -i '1,2d'
```