Fix CSV files with subscript HTML tags
When converting the creole documentation to Markdown, we substituted the
pattern `,,.*?,,` with `<sub>$1</sub>` in order to get proper subscripts
in the Markdown files. Unfortunately we also applied that
search-and-replace to CSV files which also exhibit this pattern, making
the CSV files invalid. This commit reverts that change in our data CSVs.