Convert Choices Array Separator to Pipe
In the original AMEE dataset, the last column of all the CSV files,
`options`, contains a comma-separated list of valid options for the
field in question. Since the CSV file itself is ALSO comma-separated,
that really trips up parsers and makes the files invalid to most of
them.

In order to alleviate the situation, I wrote the following script to
parse the files, get the number of columns they SHOULD have, and then
group the last n columns into an Array and join it with the `|` (Pipe)
character instead of the comma character. Here's the script, for
posterity and transparency:

```rb

require 'csv'

source_csv = CSV.read(ARGV[0], headers: true)

CSV.open(ARGV[0], 'wb') do |csv|
  csv << source_csv.headers.compact
  number_of_columns = source_csv.headers.compact.length
  source_csv.each do |line|
    new_line = line.to_a[0..(number_of_columns - 2)].map { |l| l[1] }
    choices = line.to_a[(number_of_columns -1)..-1].map{ |l| l[1] }.compact
    new_line << (choices.length > 0 ? choices.join('|') : nil)
    csv << new_line
  end
end
```

As a side-effect, this also eliminated a few extra commas at the end of
lines.